Rasberry pi + Arduino Firmware + Video
HW - RPI, arduino

Organization
Progress folder
Firmware folder
Software folder
