+++ README.md
@@ -1,3 +1,4 @@
 MTE 380 Final Project - Ball balancing Stewart platform with maze solving capabilities
 
 Code Owners: Justin Joseph, Dharma Gupta, James Li, Chanuth Weeraratna, Allen Loukiantchenko
+
